run package script in dist
