+++ tools/schematics/package.json
@@ -31,7 +31,7 @@
     "verdaccio:publish": "bash scripts/publish-local.sh",
     "verdaccio:unpublish": "bash scripts/unpublish-local.sh",
     "test": "NODE_OPTIONS=\"--no-experimental-strip-types\" ts-node --compiler-options '{\"module\":\"commonjs\"}' -P tsconfig.spec.json -r tsconfig-paths/register ../../node_modules/jasmine/bin/jasmine.js --config=jasmine.json",
-		"collect-package-versions": "tsx scripts/collect-platform-versions.mts"
+		"collect-package-versions": "node ../../dist/commerce/scripts/collect-platform-versions.js"
   },
   "dependencies": {
     "@angular-devkit/core": "0.0.0-PLACEHOLDER",
@@ -71,6 +71,11 @@
         ]
       },
       "verdaccio:publish": {
+        "dependsOn": [
+          "build"
+        ]
+      },
+			"collect-package-versions": {
         "dependsOn": [
           "build"
         ]
