+++ tools/schematics/versioning/packages.json
@@ -1,1 +1,1 @@
-{"magento":{"order":["2.4.0","2.4.1"],"external-router":["2.4.1","2.4.2","2.4.3"],"customer-order":["2.4.5","2.4.6"]}}
+{}
